fixed unfold issue and added bench.c

+++ include/splatt.h
@@ -1,7 +1,7 @@
 #ifndef SPLATT_SPLATT_H
 #define SPLATT_SPLATT_H
 
-typedef float splatt_val_t;
+typedef double splatt_val_t;
 #define SS_VAL "%f"
 typedef unsigned int splatt_idx_t;
 #define SS_IDX "%u"
